Fixes some typos, wording and formatting
While reading the documentation I found some typos I wanted to fix and also some formatting problems. 

There is also a problem with the wording of the methods: "deactive" is no word! Please rename it to inactive, which is the correct term you want to use. I'll create a issue for it.